ml_dev/pyproject.toml
[project]
name = "vhm-ml"
version = "0.0.0"
requires-python = ">=3.11"

dependencies = [
  "ipykernel",
  "matplotlib",
  "numpy",
  "torchgeo @ git+https://github.com/keves1/torchgeo.git@add-autoregression-task",
  "wandb",
  "xarray",
  "zarr",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
fix = true

[tool.ruff.lint]
extend-select = ["I", "RUF"]
